fix servicios
Ahora los servicios de usuario, pedido y notificacion siguen el estandar que decidimos usar para el lanzamiento de excepciones

+++ api/src/main/java/com/viandasApp/api/Usuario/service/UsuarioServiceImpl.java
@@ -1,9 +1,7 @@
 package com.viandasApp.api.Usuario.service;
 
 import com.viandasApp.api.Config.SecurityConfig;
-import com.viandasApp.api.Emprendimiento.model.Emprendimiento;
 import com.viandasApp.api.Emprendimiento.repository.EmprendimientoRepository;
-import com.viandasApp.api.Pedido.model.Pedido;
 import com.viandasApp.api.Pedido.repository.PedidoRepository;
 import com.viandasApp.api.Usuario.dto.UsuarioCreateDTO;
 import com.viandasApp.api.Usuario.dto.UsuarioDTO;
@@ -12,15 +10,15 @@ import com.viandasApp.api.Usuario.dto.UsuarioUpdateRolDTO;
 import com.viandasApp.api.Usuario.model.RolUsuario;
 import com.viandasApp.api.Usuario.model.Usuario;
 import com.viandasApp.api.Usuario.repository.UsuarioRepository;
-import com.viandasApp.api.Vianda.model.Vianda;
 import com.viandasApp.api.Vianda.repository.ViandaRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class UsuarioServiceImpl implements UsuarioService {
@@ -40,6 +38,7 @@ public class UsuarioServiceImpl implements UsuarioService {
         this.passwordEncoder = new SecurityConfig().passwordEncoder(); // Obtiene el PasswordEncoder de la configuración de seguridad
     }
 
+    //--------------------------Create--------------------------//
     @Transactional
     @Override
     public UsuarioDTO createUsuario(UsuarioCreateDTO usuarioCreateDTO) {
@@ -48,99 +47,128 @@ public class UsuarioServiceImpl implements UsuarioService {
         return new UsuarioDTO(savedUsuario);
     }
 
+    //--------------------------Read--------------------------//
     @Override
     public List<UsuarioDTO> readUsuarios() {
-        return usuarioRepository.findAll().stream()
-                .map(UsuarioDTO::new)
-                .collect(Collectors.toList());
+        List<UsuarioDTO> encontrados = usuarioRepository.findAll()
+                .stream()
+                .map(UsuarioDTO::new).toList();
+
+        if (encontrados.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron usuarios");
+        }
+        return encontrados;
     }
 
     @Override
     public Optional<UsuarioDTO> findById(Long id) {
-        return usuarioRepository.findById(id).map(UsuarioDTO::new);
-    }
+        Optional <UsuarioDTO> encontrado = usuarioRepository.findById(id)
+                .map(UsuarioDTO::new);
 
-    @Override
-    public Optional<Usuario> findEntityById(Long id) {
-        return usuarioRepository.findById(id);
+        if (encontrado.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron usuarios con el ID: " + id);
+        }
+
+        return encontrado;
     }
 
     @Override
     public Optional<UsuarioDTO> findByNombreCompleto(String nombreCompleto) {
-        return usuarioRepository.findByNombreCompletoContaining(nombreCompleto).map(UsuarioDTO::new);
+        Optional <UsuarioDTO> encontrado = usuarioRepository.findByNombreCompletoContaining(nombreCompleto)
+                .map(UsuarioDTO::new);
+
+        if (encontrado.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron usuarios con el nombre: " + nombreCompleto);
+        }
+
+        return encontrado;
     }
 
     @Override
     public Optional<UsuarioDTO> findByEmail(String email) {
-        return usuarioRepository.findByEmail(email)
+        Optional <UsuarioDTO> encontrado = usuarioRepository.findByEmail(email)
                 .map(UsuarioDTO::new);
+
+        if (encontrado.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron usuarios con el email: " + email);
+        }
+
+        return encontrado;
     }
 
     @Override
     public List<UsuarioDTO> findByRolUsuario(RolUsuario rolUsuario) {
-        return usuarioRepository.findByRolUsuario(rolUsuario)
+        List<UsuarioDTO> encontrados = usuarioRepository.findByRolUsuario(rolUsuario)
                 .stream()
-                .map(UsuarioDTO::new)
-                .collect(Collectors.toList());
+                .map(UsuarioDTO::new).toList();
+
+        if (encontrados.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontraron usuarios con el rol: " + rolUsuario);
+        }
+        return encontrados;
     }
 
+    //--------------------------Update--------------------------//
     @Transactional
     @Override
     public Optional<UsuarioDTO> updateUsuario(Long id, UsuarioUpdateDTO usuarioUpdateDTO, Usuario autenticado) {
         if (!autenticado.getId().equals(id)) {
-            return Optional.empty(); // acceso denegado
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acceso denegado para actualizar este usuario");
         }
 
-        return usuarioRepository.findById(id).map(usuarioExistente -> {
+        Usuario usuarioExistente = usuarioRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id));
 
-            if (usuarioUpdateDTO.getNombreCompleto() != null) {
-                usuarioExistente.setNombreCompleto(usuarioUpdateDTO.getNombreCompleto());
-            }
+        if (usuarioUpdateDTO.getNombreCompleto() != null) {
+            usuarioExistente.setNombreCompleto(usuarioUpdateDTO.getNombreCompleto());
+        }
 
-            if (usuarioUpdateDTO.getEmail() != null) {
-                usuarioExistente.setEmail(usuarioUpdateDTO.getEmail());
-            }
+        if (usuarioUpdateDTO.getEmail() != null) {
+            usuarioExistente.setEmail(usuarioUpdateDTO.getEmail());
+        }
 
-            if (usuarioUpdateDTO.getTelefono() != null) {
-                usuarioExistente.setTelefono(usuarioUpdateDTO.getTelefono());
-            }
+        if (usuarioUpdateDTO.getTelefono() != null) {
+            usuarioExistente.setTelefono(usuarioUpdateDTO.getTelefono());
+        }
 
-            Usuario actualizado = usuarioRepository.save(usuarioExistente);
-            return new UsuarioDTO(actualizado);
-        });
+        Usuario actualizado = usuarioRepository.save(usuarioExistente);
+        return Optional.of(new UsuarioDTO(actualizado));
     }
 
     @Transactional
     @Override
     public Optional<UsuarioDTO> updateUsuarioAdmin(Long id, UsuarioUpdateRolDTO usuarioUpdateRolDTO) {
-        return usuarioRepository.findById(id).map(usuarioExistente -> {
+        Usuario usuarioExistente = usuarioRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id));
 
-            if (usuarioUpdateRolDTO.getNombreCompleto() != null) {
-                usuarioExistente.setNombreCompleto(usuarioUpdateRolDTO.getNombreCompleto());
-            }
+        if (usuarioUpdateRolDTO.getNombreCompleto() != null) {
+            usuarioExistente.setNombreCompleto(usuarioUpdateRolDTO.getNombreCompleto());
+        }
 
-            if (usuarioUpdateRolDTO.getEmail() != null) {
-                usuarioExistente.setEmail(usuarioUpdateRolDTO.getEmail());
-            }
+        if (usuarioUpdateRolDTO.getEmail() != null) {
+            usuarioExistente.setEmail(usuarioUpdateRolDTO.getEmail());
+        }
 
-            if (usuarioUpdateRolDTO.getTelefono() != null) {
-                usuarioExistente.setTelefono(usuarioUpdateRolDTO.getTelefono());
-            }
+        if (usuarioUpdateRolDTO.getTelefono() != null) {
+            usuarioExistente.setTelefono(usuarioUpdateRolDTO.getTelefono());
+        }
 
-            if (usuarioUpdateRolDTO.getRolUsuario() != null) {
-                usuarioExistente.setRolUsuario(usuarioUpdateRolDTO.getRolUsuario());
-            }
+        if (usuarioUpdateRolDTO.getRolUsuario() != null) {
+            usuarioExistente.setRolUsuario(usuarioUpdateRolDTO.getRolUsuario());
+        }
 
-            Usuario actualizado = usuarioRepository.save(usuarioExistente);
-            return new UsuarioDTO(actualizado);
-        });
+        Usuario actualizado = usuarioRepository.save(usuarioExistente);
+        return Optional.of(new UsuarioDTO(actualizado));
     }
 
+    //--------------------------Delete--------------------------//
     @Transactional
     @Override
     public boolean deleteUsuarioAdmin(Long id) {
         Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
-        if (usuarioExistente.isEmpty()) return false;
+        if (usuarioExistente.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id);
+        }
 
         Usuario usuario = usuarioExistente.get();
 
@@ -151,36 +179,52 @@ public class UsuarioServiceImpl implements UsuarioService {
     @Transactional
     @Override
     public boolean deleteUsuario(Long id, Usuario autenticado) {
-        if (!autenticado.getId().equals(id)){
-            return false;
+
+        Optional<Usuario> usuarioExistente = usuarioRepository.findById(id);
+        if (usuarioExistente.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id);
         }
+
+        if (!autenticado.getId().equals(id)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acceso denegado para eliminar este usuario");
+        }
+
+        Usuario usuario = usuarioExistente.get();
+
         usuarioRepository.deleteById(id);
         return true;
     }
 
-    private Usuario DTOToEntity(UsuarioCreateDTO usuarioCreateDTO) {
+    //--------------------------Otros--------------------------//
+    private Usuario DTOToEntity(UsuarioCreateDTO dto) {
         return new Usuario(
-                usuarioCreateDTO.getId(),
-                usuarioCreateDTO.getNombreCompleto(),
-                usuarioCreateDTO.getEmail(),
-                passwordEncoder.encode(usuarioCreateDTO.getPassword()),
-                usuarioCreateDTO.getTelefono(),
-                usuarioCreateDTO.getRolUsuario()
+                null, // id será generado por JPA
+                dto.getNombreCompleto(),
+                dto.getEmail(),
+                passwordEncoder.encode(dto.getPassword()),
+                dto.getTelefono(),
+                dto.getRolUsuario()
         );
     }
 
+    @Override
+    public Optional<Usuario> findEntityById(Long id) {
+        return usuarioRepository.findById(id);
+    }
+
+    @Transactional
     @Override
     public boolean cambiarPasswordAdmin(Long id, String passwordNueva) {
         Optional<Usuario> usuarioOpt = usuarioRepository.findById(id);
 
         if (usuarioOpt.isEmpty()){
-            return false;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id);
         }
 
         Usuario usuario = usuarioOpt.get();
 
         if (passwordEncoder.matches(passwordNueva, usuario.getPassword())){
-            return false;
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "La nueva contraseña no puede ser igual a la actual");
         }
 
         usuario.setPassword(passwordEncoder.encode(passwordNueva));
@@ -188,22 +232,26 @@ public class UsuarioServiceImpl implements UsuarioService {
         return true;
     }
 
+    @Transactional
     @Override
     public boolean cambiarPassword(Long id, String passwordActual, String passwordNueva, Usuario autenticado) {
         Optional<Usuario> usuarioOpt = usuarioRepository.findById(id);
-        if (usuarioOpt.isEmpty()) return false;
+        if (usuarioOpt.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No se encontró el usuario con el ID: " + id);
+        }
+
         Usuario usuario = usuarioOpt.get();
 
         if (!usuario.getId().equals(autenticado.getId())) {
-            return false;
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Acceso denegado para cambiar la contraseña de este usuario");
         }
 
         if (!passwordEncoder.matches(passwordActual, usuario.getPassword())) {
-            return false;
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "La contraseña actual es incorrecta");
         }
 
         if (passwordEncoder.matches(passwordNueva, usuario.getPassword())) {
-            return false;
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "La nueva contraseña no puede ser igual a la actual");
         }
 
         usuario.setPassword(passwordEncoder.encode(passwordNueva));
